+++ tablesalt/lib/tablesalt/rspec/config.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.configure do |config|
+  config.around(:each) do |example|
+    Tablesalt::ThreadAccessor.clean_thread_context(&example)
+  end
+end