+++ tablesalt/lib/tablesalt/spec_helper.rb
@@ -1,3 +1,5 @@
 # frozen_string_literal: true
 
 require_relative "rspec/custom_matchers"
+require_relative "rspec/config"
+require_relative "rspec/shared_examples"