+++ tablesalt/lib/tablesalt/thread_accessor.rb
@@ -2,9 +2,17 @@
 
 require "active_support/concern"
 
+require_relative "thread_accessor/management"
+require_relative "thread_accessor/rack_middleware"
+require_relative "thread_accessor/thread_store"
+
+# WARNING: This module is still in beta mode and will likely change significantly soon. Tread carefully...
 module Tablesalt
   module ThreadAccessor
     extend ActiveSupport::Concern
+    extend Management
+
+    STORE_THREAD_KEY = :__tablesalt_thread_accessor_store__
 
     module ClassMethods
       private
@@ -29,8 +37,8 @@ module Tablesalt
       # @param thread_key [String, Symbol] The key to read from Thread.current
       # @option :private [Boolean] If true, both defined methods will be private. Default: true
       def thread_reader(method, thread_key, **options)
-        define_method(method) { Thread.current[thread_key] }
-        define_singleton_method(method) { Thread.current[thread_key] }
+        define_method(method) { ThreadAccessor.store[thread_key] }
+        define_singleton_method(method) { ThreadAccessor.store[thread_key] }
 
         return unless options.fetch(:private, true)
 
@@ -52,14 +60,20 @@ module Tablesalt
       #   Thread.current[:foo]
       #   => "bar"
       #
+      # Note::
+      #   All written thread variables are tracked on-thread, but will not be automatically cleared when
+      #   the thread is done processing a request/unit of work. Make sure to either use the provided
+      #   {ThreadAccessor::RackMiddleware} or run {ThreadAccessor.clear_thread_variables!} manually once
+      #   the thread is finished to avoid pollluting other requests.
+      #
       # @param method [String, Symbol] The name of the writer method
       # @param thread_key [String, Symbol] The key to write to on Thread.current
       # @option :private [Boolean] If true, both defined methods will be private. Default: true
       def thread_writer(method, thread_key, **options)
         method_name = "#{method}="
 
-        define_method(method_name) { |value| Thread.current[thread_key] = value }
-        define_singleton_method(method_name) { |value| Thread.current[thread_key] = value }
+        define_method(method_name) { |value| ThreadAccessor.store[thread_key] = value }
+        define_singleton_method(method_name) { |value| ThreadAccessor.store[thread_key] = value }
 
         return unless options.fetch(:private, true)